Finished everything up. Fixed some of my error codes and making the double save two decimal points. I forgot that dividing interger values will give an interger output so I fixed that.

+++ prepare/Learning03/Fraction.cs
@@ -35,7 +35,7 @@ class Fraction
     
     public double GetFraction()
     {
-        double _fraction = (_top/_bottom);
+        double _fraction = (double)_top/_bottom;
         return _fraction;
     }
 }
